refactor: move/rename header menu item and dropdown
